notmuchsync: make Message.parse() not output scary warnings on an empty text
When notmuch show does not return any results we would try to hand Message.parse() an empty text. So far that lead to the 'no valid mail' output, but this was not really an error. So just log.debug() something and return without error if we hand it an empty string.

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>