1)Reunify the 'notmuch' module into the notmuchsync script. 2)Proper warnings on import errors
The backend module is now cnotmuch, so there is no need to keep that minimal notmuch module around.

As this version requires the external cnotmuch bindings and a libnotmuch.so as of new, we also check
if the import of cnotmuch went fine and if we could load the shared library. Bail out with a
somewhat friendly error message if loading cnotmuch failed.

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>