don't sync "deleted" by default
as the thread around mail 877hie4o1w.fsf@raven.defaultvalue.org discusses,
can notmuchsync not always be sure that synching the "deleted"
flag modifies the correct message (as there might be multiple copies
with the same message id), so synching the "deleted" tag (although
usually correct), is a bit more dangerous than the rest of the tags.
Disable synching the "deleted" tag and "trashed" maildir flag by default
and only enable it with an explicit --sync-deleted option.